refactor: split command_contract output routing from output type definitions
command_contract/output.rs mixed the output-contract TYPES (CommandDescriptor,
CommandResponsePlan, the output/response enums) with an impl Commands block that
inspects the parsed CLI command enum to derive per-command output descriptors.
The impl pulls in cli_surface::Commands and eight command handler modules,
forcing a commands dependency onto the output type file.

Move the impl Commands routing (plus its private helpers and the descriptor-
drift test) into a new command_contract/output_routing.rs, mirroring the earlier
lab.rs/lab_routing.rs split. output.rs now holds only type definitions and the
impl CommandOutputDescriptor bridge (with_lab_contract promoted to pub(crate) so
the relocated routing can call it).

output.rs is now free of commands/cli_surface. The remaining command_contract
commands-coupling is isolated in the dedicated dispatch/routing modules
(descriptors macro, lab_routing, output_routing) plus safety_manifest's
cli_surface type imports -- the next step relocates those fully to the CLI side.

No behavior change. cargo check --lib and --lib --tests pass.
